Use same dtype tensors as detector_response

+++ flamedisx/core.py
@@ -418,13 +418,13 @@ class ERSource:
         """
         n_det, n_prod = self.cross_domains(quanta_type + '_detected',
                                            quanta_type + '_produced')
-        n_det = tf.convert_to_tensor(n_det, dtype=tf.float32)
-        n_prod = tf.convert_to_tensor(n_prod, dtype=tf.float32)
+        n_det = tf.convert_to_tensor(n_det, dtype=tf.float64)
+        n_prod = tf.convert_to_tensor(n_prod, dtype=tf.float64)
         p = self.gimme(quanta_type + '_detection_eff')[:, o, o]
         if quanta_type == 'photon':
             # Note *= doesn't work, p will get reshaped
             p = p * self.gimme('penning_quenching_eff', n_prod)
-        p = tf.convert_to_tensor(p, dtype=tf.float32)
+        p = tf.convert_to_tensor(p, dtype=tf.float64)
         result = tfd.Binomial(total_count=n_prod, probs=p).prob(n_det)
         return result * self.gimme(quanta_type + '_acceptance', n_det)
